requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
pydantic==2.11.1
psycopg2-binary==2.9.10
python-dotenv==1.2.2
